fixed tg_ppo_agent.py to work also with torch (moved functions to policy model)
