+++ src/agents/ppo_agent.py
@@ -254,12 +254,12 @@ class PPOAgent:
         state_batch = torch.tensor(self.batch_states, device=self.device, dtype=torch.float32)
         # batch_state_values = self.policy.v(state_batch).view(-1, 1)
         batch_state_values = torch.tensor(self.batch_vals, device=self.device, dtype=torch.float32).view(-1, 1)
-        batch_advantage_tensor = self.normalize_batch_advantage(batch_rtgs_tensor - batch_state_values.detach())
         # Policy loss
         pi, chosen_logprob = self.policy.pi(state_batch.clone(), torch.tensor(self.batch_actions).to(device=self.device))
         chosen_logprob = chosen_logprob.view(-1, 1)
-        ratio = torch.exp(chosen_logprob - original_logprobs)
 
+        batch_advantage_tensor = self.normalize_batch_advantage(batch_rtgs_tensor - batch_state_values.detach())
+        ratio = torch.exp(chosen_logprob - original_logprobs)
         clip_adv = torch.clamp(ratio, 1 - self.eps_clip, 1 + self.eps_clip) * batch_advantage_tensor
         policy_loss = -(torch.min(ratio * batch_advantage_tensor, clip_adv)).mean()
